Moving score to models
