+++ pyrit/models/__init__.py
@@ -15,6 +15,7 @@ from pyrit.models.chat_message import ChatMessage, ChatMessageListContent
 from pyrit.models.prompt_request_piece import PromptRequestPiece
 from pyrit.models.prompt_request_response import PromptRequestResponse, group_conversation_request_pieces_by_sequence
 from pyrit.models.identifiers import Identifier
+from pyrit.models.score import Score, ScoreType
 
 
 __all__ = [
