+++ pyrit/score/self_ask_category_scorer.py
@@ -133,7 +133,7 @@ class SelfAskCategoryScorer(Scorer):
                 score_category=parsed_response["category_name"],
                 score_rationale=parsed_response["rationale"],
                 scorer_class_identifier=self.get_identifier(),
-                metadata=None,
+                score_metadata=None,
                 prompt_request_response_id=request_response.id,
             )
             return [score]
