+++ pyrit/score/self_ask_likert_scorer.py
@@ -132,7 +132,7 @@ class SelfAskLikertScorer(Scorer):
                 score_category=self._score_category,
                 score_rationale=parsed_response["rationale"],
                 scorer_class_identifier=self.get_identifier(),
-                metadata=None,
+                score_metadata=None,
                 prompt_request_response_id=request_response.id,
             )
             return [score]
